+++ NeuralNetworkCode/src/Stimulus/WhiteNoiseLinear.cpp
@@ -28,16 +28,16 @@ void WhiteNoiseLinear::LoadParameters(std::vector<std::string> *input) {
 		SplitString(&(*it), &name, &values);
 
 		if ((name.find("seed") != std::string::npos)) {
-			seed = std::stod(values.at(0));
+			seed = static_cast<unsigned int>(std::stod(values.at(0)));
 		}
 		else if ((name.find("meanCurrent") != std::string::npos)) {
 			s.values.clear();
 			for (unsigned int i = 0;i < P;i++) {
-				s.values.push_back(std::stod(values.at(2 * i)));
-				s.values.push_back(std::stod(values.at(2 * i + 1)));
+				s.values.push_back(std::stod(values.at(2 * static_cast<long long>(i))));
+				s.values.push_back(std::stod(values.at(2 * static_cast<long long>(i) + 1)));
 			}
-			s.start_time = std::round(std::stod(values.at(2 * P)) / info->dt);
-			s.end_time   = std::round(std::stod(values.at(2 * P + 1)) / info->dt);
+			s.start_time = static_cast<int>(std::round(std::stod(values.at(2 * P)) / info->dt));
+			s.end_time = static_cast<int>(std::round(std::stod(values.at(2 * P + 1)) / info->dt));
 
 			meanCurrent.push_back(s);
 		}
@@ -47,8 +47,8 @@ void WhiteNoiseLinear::LoadParameters(std::vector<std::string> *input) {
 				s.values.push_back(std::stod(values.at(2*i)));
 				s.values.push_back(std::stod(values.at(2*i+1)));
 				}
-			s.start_time = std::round(std::stod(values.at(2 * P)) / info->dt);
-			s.end_time   = std::round(std::stod(values.at(2 * P + 1)) / info->dt);
+			s.start_time = static_cast<int>(std::round(std::stod(values.at(2 * P)) / info->dt));
+			s.end_time = static_cast<int>(std::round(std::stod(values.at(2 * P + 1)) / info->dt));
 			sigmaCurrent.push_back(s);
 		}
 	}
@@ -75,8 +75,8 @@ void WhiteNoiseLinear::SaveParameters(std::ofstream * stream) {
 			*stream << std::to_string(s.values.at(2 * i)) << "\t ";
 			*stream << std::to_string(s.values.at(2 * i + 1)) << "\t ";
 		}
-		*stream << std::to_string(((double)s.start_time)*info->dt) << " \t";
-		*stream << std::to_string(((double)s.end_time)*info->dt) << " \t";
+		*stream << std::to_string(static_cast<double>(s.start_time)*info->dt) << " \t";
+		*stream << std::to_string(static_cast<double>(s.end_time)*info->dt) << " \t";
 		*stream << " [col 1: input to pop0 at t_0, col 2: pop1 at t_0, ... colP+1: pop1 t_f, ... col2P: popN t_f, t0, tf. Dimensions: [mV/sec , secs.]\n";
 	}
 
@@ -86,8 +86,8 @@ void WhiteNoiseLinear::SaveParameters(std::ofstream * stream) {
 			*stream << std::to_string(s.values.at(2 * i)) << "\t ";
 			*stream << std::to_string(s.values.at(2 * i + 1)) << "\t ";
 		}
-		*stream << std::to_string(((double)s.start_time)*info->dt) << " \t";
-		*stream << std::to_string(((double)s.end_time)*info->dt) << " \t";
+		*stream << std::to_string(static_cast<double>(s.start_time)*info->dt) << " \t";
+		*stream << std::to_string(static_cast<double>(s.end_time)*info->dt) << " \t";
 		*stream << " [col 1: input to pop0 at t_0, col 2: pop1 at t_0, ... colP+1: pop1 t_f, ... col2P: popN t_f, t0, tf. Dimensions: [mV/sqrt(sec) , secs.]\n";
 	}
